Version 10.0
Cycle through all the character ranges specified in the
iskeyword option and build a list of valid word separators.

Prior to this change, only actual characters were used,
where for example ASCII "45" == "-".  If "45" were used
in iskeyword the hyphen would not be picked up.

This introduces a new option, since the character ranges
specified could be multibyte:
    let g:omni_syntax_use_single_byte = 1

This by default will only allow single byte ASCII
characters to be added and an additional check to ensure
the character is printable (see documentation for isprint).